Split display-specific function into a separate module
